@@ -1,8 +1,7 @@
 package ru.androidclass.easybackupsample.db.dao;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Query;
+import androidx.room.Dao;
+import androidx.room.Query;
 
 import java.util.List;
 
